requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.9
cloudinary==1.28.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.8.0
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djrill==2.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
Pillow==8.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-http-client==3.3.4
pytz==2021.3
requests==2.26.0
sendgrid==3.6.5
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
zope.interface==5.4.0
